+++ plug-api/plugos-syscall/mod.ts
@@ -2,7 +2,6 @@ export * as asset from "./asset.ts";
 export * as events from "./event.ts";
 // export * as fs from "./fs.ts";
 export { LocalFileSystem } from "./fs.ts";
-export * as fulltext from "./fulltext.ts";
 export * as shell from "./shell.ts";
 export * as store from "./store.ts";
 export * as YAML from "./yaml.ts";
