+++ plugs/core/core.plug.yaml
@@ -136,40 +136,7 @@ functions:
     path: "./anchor.ts:anchorComplete"
     events:
       - editor:complete
-
-  # Full text search
-  # searchIndex:
-  #   path: ./search.ts:pageIndex
-  #   events:
-  #     - page:index
-  # searchUnindex:
-  #   path: "./search.ts:pageUnindex"
-  #   env: server
-  #   events:
-  #     - page:deleted
-  # searchQueryProvider:
-  #   path: ./search.ts:queryProvider
-  #   events:
-  #     - query:full-text
-  # searchCommand:
-  #   path: ./search.ts:searchCommand
-  #   command:
-  #     name: "Search Space"
-  #     key: Ctrl-Shift-f
-  #     mac: Cmd-Shift-f
-  # readPageSearch:
-  #   path: ./search.ts:readFileSearch
-  #   env: server
-  #   pageNamespace:
-  #     pattern: "🔍 .+"
-  #     operation: readFile
-  # getPageMetaSearch:
-  #   path: ./search.ts:getFileMetaSearch
-  #   env: server
-  #   pageNamespace:
-  #     pattern: "🔍 .+"
-  #     operation: getFileMeta
-
+  
   # Template commands
   insertTemplateText:
     path: "./template.ts:insertTemplateText"
