+++ plugs/search/search.plug.yaml
@@ -4,10 +4,43 @@ functions:
     path: search.ts:indexPage
     events:
     - page:index
+
   clearIndex:
     path: search.ts:clearIndex
 
-  test:
-    path: search.ts:testCommand
+  searchUnindex:
+    path: "./search.ts:pageUnindex"
+    env: server
+    events:
+      - page:deleted
+  searchQueryProvider:
+    path: ./search.ts:queryProvider
+    events:
+      - query:full-text
+  searchCommand:
+    path: ./search.ts:searchCommand
     command:
-      name: "Search: Test"
+      name: "Search Space"
+      key: Ctrl-Shift-f
+      mac: Cmd-Shift-f
+
+  readPageSearch:
+    path: ./search.ts:readFileSearch
+    env: server
+    pageNamespace:
+      pattern: "🔍 .+"
+      operation: readFile
+
+  writePageSearch:
+    path: ./search.ts:writeFileSearch
+    env: server
+    pageNamespace:
+      pattern: "🔍 .+"
+      operation: readFile
+
+  getPageMetaSearch:
+    path: ./search.ts:getFileMetaSearch
+    env: server
+    pageNamespace:
+      pattern: "🔍 .+"
+      operation: getFileMeta
